TaskService/requirements.txt
Flask==3.0.2
flask_sqlalchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
psycopg2==2.9.9
grpcio==1.51.1
grpcio-tools==1.51.1